// app/types/accountManagement.ts
export type ProfileFormValues = {
   avatar?: string
   username?: string
   email?: string
   registrationDate?: string
}

export type AffiliateFormValues = {
   affiliatePasscode?: string
   accountBank: string
   bank: string
   city: string
   branch: string
   bankName: string
}

export type PasswordFormValues = {
   currentPassword: string
   password: string
   confirmPassword: string
}